+++ patchwork-registries/src/main/resources/patchwork-registries.mixins.json
@@ -12,8 +12,6 @@
     "MixinChunkGeneratorType",
     "MixinChunkStatus",
     "MixinContainerType",
-    "MixinCookingRecipeSerializer",
-    "MixinCuttingRecipeSerializer",
     "MixinDecorator",
     "MixinEnchantment",
     "MixinEntityType",
@@ -26,12 +24,10 @@
     "MixinPointOfInterestType",
     "MixinPotion",
     "MixinRecipeSerializer",
+    "MixinRecipeSerializerSubclass",
     "MixinSchedule",
     "MixinSensorType",
-    "MixinShapedRecipeSerializer",
-    "MixinShapelessRecipeSerializer",
     "MixinSoundEvent",
-    "MixinSpecialRecipeSerializer",
     "MixinStatType",
     "MixinStatusEffect",
     "MixinSurfaceBuilder",
